requirements.txt
numpy>=1.21.0
scipy
scikit-learn
numba>=0.55.0rc1
cython>=0.21.0
pandas
tqdm
torch>=1.7.0
torchvision
torchaudio
transformers
hydra-core
pysptk
pyworld
tensorboard
pandas
matplotlib
soundfile
faster_whisper
nnmnkwii
g2p_en
gradio<=4.21.0
